scraper/requirements.txt
sqlalchemy~=1.4.27
aiofiles~=0.8.0
APScheduler~=3.8.1
psycopg2
sqlalchemy_utils
pyyaml
iso8601
python-dateutil
lxml>=2.2
pytz
cssselect
scrapelib
mechanize
BeautifulSoup4
mock
xmltodict
rtyaml